feat: fietsafstand werklocaties + codebase cleanup
Fietsafstand feature (closes #42):
- OpenRouteService cycling collector met caching (7 dagen)
- Werklocaties config (CS Den Haag, Mariahoeve) in config/werklocaties.yaml
- Fietsafstanden geintegreerd in /api/voorzieningen response
- VoorzieningenPanel: fietsafstand sectie met kleurcodes + route polylines op kaart
- ORS_API_KEY in .env.example

Codebase cleanup:
- woningen.py (1157 regels) gesplitst in woningen.py (listings) + waardebepaling.py (valuation)
- Gedeelde address parsing utility (backend/utils/address.py) vervangt 3x duplicatie
- Bug fix: fallback huisnummer=1 bij parse-fout -> geeft nu HTTPException
- Bug fix: get_session import in voorzieningen.py -> SessionLocal()
- GET/POST WOZ en energielabel geconsolideerd via _lookup_woz()/_lookup_energielabel()
- Duplicate httpx verwijderd uit requirements.txt
- Hardcoded TODO-velden nieuw_deze_week/verkocht_deze_week verwijderd uit MarktOverzicht
