+++ backend/.env.example
@@ -8,3 +8,7 @@ BAG_API_KEY=your_bag_api_key_here
 # EP-Online API Key (RVO - Energielabels)
 # Request your API key at: https://www.rvo.nl/onderwerpen/energielabel-aanvragen
 EP_ONLINE_API_KEY=your_ep_online_api_key_here
+
+# OpenRouteService API Key (cycling routes)
+# Get your free API key at: https://openrouteservice.org/dev/#/signup
+ORS_API_KEY=your_ors_api_key_here
