+++ backend/api/__init__.py
@@ -2,6 +2,7 @@
 
 from api.buurten import router as buurten_router
 from api.woningen import router as woningen_router
+from api.waardebepaling import router as waardebepaling_router
 from api.watchlist import router as watchlist_router
 from api.markt import router as markt_router
 from api.scholen import router as scholen_router
@@ -10,6 +11,7 @@ from api.voorzieningen import router as voorzieningen_router
 __all__ = [
     "buurten_router",
     "woningen_router",
+    "waardebepaling_router",
     "watchlist_router",
     "markt_router",
     "scholen_router",
