+++ backend/utils/__init__.py
@@ -1,0 +1,5 @@
+"""Utility modules for Woningzoeker backend."""
+
+from utils.address import parse_huisnummer
+
+__all__ = ["parse_huisnummer"]
