+++ requirements.txt
@@ -26,4 +26,3 @@ pyproj>=3.6.0
 # Dev dependencies
 pytest>=7.4.0
 pytest-asyncio>=0.23.0
-httpx>=0.26.0
